requirements.txt
aqa-sitronics==0.6
flake8-builtins==1.5.3
flake8-commas==2.1.0
flake8-function-order==0.0.5
flake8-functions==0.0.7
flake8-import-order==0.18.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pytest==7.1.2
requests==2.27.1
selenium==4.2.0
faker
